reorganised cli package functionalities
By reoganising the cli package i isolated the function (sendAnalytics) i will work on in the next PR.
